fix : some functions moved to orangetool_utils.py
